+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Add `iosxe_dot1x` resource and data source
 - Add `arp_timeout`, `spanning_tree_link_type` and `spanning_tree_portfast_trunk` attribute to `iosxe_interface_ethernet` resource and data source
 - Add `message_digest_keys` attribute to `iosxe_interface_ospf` resource and data source
+- Add `areas` and `passive_interface_default` attribute to `iosxe_ospf_vrf`  resources and data sources
 
 ## 0.3.3